Final2

+++ src/main/java/com/example/demo/controller/LoanPageController.java
@@ -2,11 +2,14 @@ package com.example.demo.controller;
 
 import com.example.demo.dto.LoanRequestDTO;
 import com.example.demo.entity.Loan;
+import com.example.demo.entity.User;
 import com.example.demo.service.AssetService;
 import com.example.demo.service.LoanService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
 @RequestMapping("/admin/loans")
@@ -24,8 +27,13 @@ public class LoanPageController {
     }
 
     @GetMapping
-    public String showLoanPage(Model model) {
+    public String showLoanPage(Model model, HttpSession session) {
 
+        User user = (User) session.getAttribute("user");
+
+        if (user == null || user.getRole() != User.Role.ADMIN) {
+            return "redirect:/loginpage";
+        }
         model.addAttribute(
                 "assets",
                 assetService.getAvailableAssets()
@@ -50,7 +58,13 @@ public class LoanPageController {
     @PostMapping("/request")
     public String requestLoan(
             @RequestParam Long userId,
-            @RequestParam Long assetId) {
+            @RequestParam Long assetId,
+            HttpSession session) {
+
+        User user = (User) session.getAttribute("user");
+        if (user == null || user.getRole() != User.Role.ADMIN) {
+            return "redirect:/loginpage";
+        }
 
         LoanRequestDTO dto = new LoanRequestDTO();
 
@@ -62,38 +76,76 @@ public class LoanPageController {
         return "redirect:/admin/loans";
     }
 
+
     @PostMapping("/return/{id}")
-    public String returnLoan(@PathVariable Long id) {
+    public String returnLoan(@PathVariable Long id, HttpSession session, RedirectAttributes redirectAttributes) {
+        User user = (User) session.getAttribute("user");
+
+        if (user == null || user.getRole() != User.Role.ADMIN) {
+            return "redirect:/loginpage";
+        }
+
+        redirectAttributes.addFlashAttribute(
+                "adminLoanPageSuccessMessage",
+                "Asset returned successfully."
+        );
         loanService.returnLoan(id);
+
         return "redirect:/admin/loans";
     }
 
-    //MODIFIED AND WORKS
     @PostMapping("/approve/{id}")
     public String approveLoan(
-            @PathVariable Long id) {
+            @PathVariable Long id, HttpSession session, RedirectAttributes redirectAttributes) {
 
+        User user = (User) session.getAttribute("user");
+
+        if (user == null || user.getRole() != User.Role.ADMIN) {
+            return "redirect:/loginpage";
+        }
         loanService.approveLoan(id);
+
+        redirectAttributes.addFlashAttribute(
+                "adminLoanPageSuccessMessage",
+                "Loan approved successfully."
+        );
 
         return "redirect:/admin/loans";
     }
 
     @PostMapping("/reject/{id}")
     public String rejectLoan(
-            @PathVariable Long id) {
+            @PathVariable Long id, HttpSession session, RedirectAttributes redirectAttributes) {
+        User user = (User) session.getAttribute("user");
 
+        if (user == null || user.getRole() != User.Role.ADMIN) {
+            return "redirect:/loginpage";
+        }
         loanService.rejectLoan(id);
+
+        redirectAttributes.addFlashAttribute(
+                "adminLoanPageSuccessMessage",
+                "Loan rejected successfully."
+        );
 
         return "redirect:/admin/loans";
     }
 
     @PostMapping("/delete/{id}")
     public String deleteLoan(
-            @PathVariable Long id) {
+            @PathVariable Long id, HttpSession session) {
+        User user = (User) session.getAttribute("user");
 
+        if (user == null || user.getRole() != User.Role.ADMIN) {
+            return "redirect:/loginpage";
+        }
         loanService.deleteLoan(id);
 
         return "redirect:/admin/loans";
     }
 
+
+
 }
+
+
